requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
gspread==3.6.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PySocks==1.7.1
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tweepy==3.9.0
urllib3==1.25.10
Werkzeug==1.0.1
